requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
babel==2.16.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==43.0.3
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
durationpy==0.9
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gensim==4.3.3
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.8
importlib_metadata==8.5.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaitaistruct==0.10
kubernetes==31.0.0
langchain==0.3.14
langchain-chroma==0.2.0
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langsmith==0.2.10
lazy-object-proxy==1.10.0
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.59.5
openai-whisper==20240930
openapi-core==0.19.4
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.14
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pathable==0.4.3
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
portalocker==2.10.1
posthog==3.7.5
prance==23.6.21.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybars4==0.9.13
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydivert==2.1.0
Pygments==2.18.0
PyJWT==2.9.0
PyMeta3==0.5.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scipy==1.13.1
selenium==4.25.0
selenium-wire==5.1.0
semantic-kernel==1.14.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.2
tbb==2021.13.1
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.20.1
torch==2.3.0
torchtext==0.18.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.1
types-python-dateutil==2.9.0.20240821
typing-inspect==0.9.0
typing_extensions==4.12.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
watchfiles==1.0.3
wcwidth==0.2.13
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
whisper==1.1.10
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.17.1
yt-dlp==2024.10.22
zipp==3.21.0
zstandard==0.23.0
